Removal of FILEPATH token
The FILENAME token conflicts with every fragment library written yet.
As such it needs to be rewritten as a parser rule w/ token rewrites.
Removing the FILEPATH token, which relies on it is the first step in doing so.
This commit rewrites the FILEPATH token as a parser rule w/ rewrite rules.