Increase ANTLR lookahead timeout in test script
The grammar is much too complex to run on some slower computers.
Yields an error that creating the DFA lookahead tree timed out in 10000 ms.
This fix updates the time before timeout to 20000 ms.

Error occurs on a server running an Opteron dual core clocked at 2Ghz.
This is the server running the Hudson CI system, thus causing build fails.