Double/single bracket conditional differences
Bash treats single bracket/test ([, test) and double bracket ([[) differently.
Certain characters in the single bracket conditional need to be escaped.
AND and OR are different as well: single bracket and test use -a and -o.
Adds support to the parser to handle the physical differences between these.
More support (lack of glob splitting, etc) will be handled in the runtime.
Tests using the new test format ([[) are marked with KEYWORD_TEST in the tree.
Tests using the old test format ([, test) are marked with BUILTIN_TEST.