Handle autoreconf warnings
Enable warnings in autogen.sh by passing -W all to autoreconf. This
pointed out things that are now fixed. AC_CHECK_FILE assumes you are
checking for a runtime file so it fails for cross compiling and as such
is not for our usage. Instead configure.ac now uses test -f.